requirements.txt
asgiref==3.10.0
Django==5.2.7
sqlparse==0.5.3
tzdata==2025.2

python-dotenv==1.0.0
google-generativeai==0.8.5
requests==2.31.0
